Fix save as draft guard when sending registration Fix refresh my observations on sync items complete

+++ src/app/modules/registration/pages/save-as-draft.guard.ts
@@ -6,6 +6,7 @@ import { RegistrationService } from '../services/registration.service';
 import { UserSettingService } from '../../../core/services/user-setting/user-setting.service';
 import { ObsLocationPage } from './obs-location/obs-location.page';
 import { TranslateService } from '@ngx-translate/core';
+import { RegistrationStatus } from '../models/registrationStatus.enum';
 
 @Injectable()
 export class SaveAsDraftRouteGuard implements CanDeactivate<OverviewPage | ObsLocationPage> {
@@ -21,31 +22,34 @@ export class SaveAsDraftRouteGuard implements CanDeactivate<OverviewPage | ObsLo
         __: RouterStateSnapshot,
         nextState?: RouterStateSnapshot) {
         if (nextState && nextState.url.indexOf('registration/') < 0 && component.registration) {
-            const translations = await this.translateService.get([
-                'REGISTRATION.SAVE_ALERT.HEADER',
-                'REGISTRATION.SAVE_ALERT.MESSAGE',
-                'REGISTRATION.SAVE_ALERT.NO',
-                'REGISTRATION.SAVE_ALERT.YES',
-            ]).toPromise();
-            const alert = await this.alertController.create({
-                header: translations['REGISTRATION.SAVE_ALERT.HEADER'],
-                message: translations['REGISTRATION.SAVE_ALERT.MESSAGE'],
-                buttons: [
-                    {
-                        text: translations['REGISTRATION.SAVE_ALERT.NO'],
-                        role: 'cancel'
-                    },
-                    {
-                        text: translations['REGISTRATION.SAVE_ALERT.YES'],
-                    }
-                ]
-            });
-            alert.present();
-            const result = await alert.onDidDismiss();
-            const save: boolean = result.role !== 'cancel';
-            if (!save) {
-                const userSettings = await this.userSettingService.getUserSettings();
-                await this.registrationService.deleteRegistrationById(userSettings.appMode, component.registration.id);
+            const reg = await this.registrationService.getSavedRegistrationById(component.registration.id);
+            if (reg.status === RegistrationStatus.Draft) {
+                const translations = await this.translateService.get([
+                    'REGISTRATION.SAVE_ALERT.HEADER',
+                    'REGISTRATION.SAVE_ALERT.MESSAGE',
+                    'REGISTRATION.SAVE_ALERT.NO',
+                    'REGISTRATION.SAVE_ALERT.YES',
+                ]).toPromise();
+                const alert = await this.alertController.create({
+                    header: translations['REGISTRATION.SAVE_ALERT.HEADER'],
+                    message: translations['REGISTRATION.SAVE_ALERT.MESSAGE'],
+                    buttons: [
+                        {
+                            text: translations['REGISTRATION.SAVE_ALERT.NO'],
+                            role: 'cancel'
+                        },
+                        {
+                            text: translations['REGISTRATION.SAVE_ALERT.YES'],
+                        }
+                    ]
+                });
+                alert.present();
+                const result = await alert.onDidDismiss();
+                const save: boolean = result.role !== 'cancel';
+                if (!save) {
+                    const userSettings = await this.userSettingService.getUserSettings();
+                    await this.registrationService.deleteRegistrationById(userSettings.appMode, component.registration.id);
+                }
             }
         }
         return true;
